refactor: Extract a module for the parsed DirsAndFileName structure
